// src/app/games/anno1800/game/enums.ts
export enum Good {
  Unknown = 'Unknown',
  AdvancedWeapons = 'Advanced Weapons',
  AlpacaWool = 'Alpaca Wool',
  AluminiumProfiles = 'Aluminium Profiles',
  ArcticGas = 'Arctic Gas',
  Bauxite = 'Bauxite',
  BearFur = 'Bear Fur',
  Beef = 'Beef',
  Beer = 'Beer',
  Beeswax = 'Beeswax',
  BilliardTables = 'Billiard Tables',
  Biscuits = 'Biscuits',
  Bombs = 'Bombs',
  BowlerHats = 'Bowler Hats',
  Brass = 'Brass',
  Bread = 'Bread',
  Bricks = 'Bricks',
  Calamari = 'Calamari',
  CamphorWax = 'Camphor Wax',
  CannedFood = 'Canned Food',
  Caoutchouc = 'Caoutchouc',
  CaribouMeat = 'Caribou Meat',
  Celluloid = 'Celluloid',
  Cement = 'Cement',
  Ceramics = 'Cermaics',
  Champagne = 'Champagne',
  CherryWood = 'Cherry Wood',
  Chocolate = 'Chocolate',
  Cigars = 'Cigars',
  Cinnamon = 'Cinnamon',
  Citrus = 'Citrus',
  Clay = 'Clay',
  ClayPipes = 'Clay Pipes',
  Chassis = 'Chassis',
  Coal = 'Coal',
  Cocoa = 'Cocoa',
  CoconutOil = 'Coconut Oil',
  Coffee = 'Coffee',
  CoffeeBeans = 'Coffee Beans',
  Cognac = 'Cognac',
  Copper = 'Copper',
  Corn = 'Corn',
  Costumes = 'Costumes',
  Cotton = 'Cotton',
  CottonFabric = 'Cotton Fabric',
  DriedMeat = 'Dried Meat',
  Dung = 'Dung',
  Dynamite = 'Dynamite',
  ElectricCables = 'Electric Cables',
  Elevators = 'Elevators',
  Ethanol = 'Ethanol',
  Fans = 'Fans',
  Felt = 'Felt',
  Fertiliser = 'Fertiliser',
  Filaments = 'Filaments',
  FilmReel = 'Film Reel',
  Finery = 'Finery',
  FireExtinguishers = 'Fire Extinguishers',
  Fish = 'Fish',
  FishOil = 'Fish Oil',
  Flour = 'Flour',
  FriedPlantains = 'Fried Plantains',
  FurCoats = 'Fur Coats',
  Furs = 'Furs',
  Glass = 'Glass',
  Glasses = 'Glasses',
  GoatMilk = 'Goat Milk',
  Gold = 'Gold',
  GoldOre = 'Gold Ore',
  GooseFeathers = 'Goose Feathers',
  Goulash = 'Goulash',
  Grain = 'Grain',
  Grapes = 'Grapes',
  Gramophones = 'Gramophones',
  Helium = 'Helium',
  Herbs = 'Herbs',
  HibiscusPetals = 'Hibiscus Petals',
  HibiscusTea = 'Hibiscus Tea',
  Hops = 'Hops',
  Huskies = 'Huskies',
  HuskySleds = 'Husky Sleds',
  IceCream = 'Ice Cream',
  IlluminatedScript = 'Illuminated Script',
  IndigoDye = 'Indigo Dye',
  IndustrialLubricant = 'Industrial Lubricant',
  Iron = 'Iron',
  Jalea = 'Jalea',
  Jam = 'Jam',
  Jewellery = 'Jewellery',
  Lacquer = 'Lacquer',
  Lanterns = 'Lanterns',
  LeatherBoots = 'Leather Boots',
  Lemonade = 'Lemonade',
  LightBulbs = 'Light Bulbs',
  Linen = 'Linen',
  Linseed = 'Linseed',
  Lobster = 'Lobster',
  Malt = 'Malt',
  Medicine = 'Medicine',
  Mezcal = 'Mezcal',
  Milk = 'Milk',
  Minerals = 'Minerals',
  Motor = 'Motor',
  MudBricks = 'Mud Bricks',
  NanduFeathers = 'Nandu Feathers',
  NanduLeather = 'Nandu Leather',
  Oil = 'Oil',
  OilLamps = 'Oil Lamps',
  Orchid = 'Orchid',
  OrnateCandles = 'Ornate Candles',
  Pamphlets = 'Pamphlets',
  Paper = 'Paper',
  Parkas = 'Parkas',
  Pearls = 'Pearls',
  Pemmican = 'Pemmican',
  PennyFarthings = 'Penny Farthings',
  Perfumes = 'Perfumes',
  Pigments = 'Pigments',
  Pigs = 'Pigs',
  Plantains = 'Plantains',
  PocketWatches = 'Pocket Watches',
  PoliceEquipment = 'Police Equipment',
  Ponchos = 'Ponchos',
  Potatoes = 'Potatoes',
  QuartzSand = 'Quartz Sand',
  RedPeppers = 'Red Peppers',
  ReinforcedConcrete = 'Reinforced Concrete',
  Resin = 'Resin',
  Rum = 'Rum',
  Sails = 'Sails',
  Salt = 'Salt',
  Saltpetre = 'Saltpetre',
  SangaCow = 'Sanga Cow',
  Sausages = 'Sausages',
  Schnapps = 'Schnapps',
  Scooter = 'Scooter',
  SeafoodStew = 'Seafood Stew',
  SealSkin = 'Seal Skin',
  SewingMachines = 'Sewing Machines',
  Shampoo = 'Shampoo',
  Sleds = 'Sleds',
  SleepingBags = 'Sleeping Bags',
  Soap = 'Soap',
  SoccerBalls = 'Soccer Balls',
  Souvenirs = 'Souvenirs',
  SpicedFlour = 'Spiced Flour',
  Spices = 'Spices',
  SteamCarriages = 'Steam Carriages',
  SteamMotors = 'Steam Motors',
  Steel = 'Steel',
  SteelBeams = 'Steel Beams',
  Sugar = 'Sugar',
  SugarCane = 'Sugar Cane',
  TailoredSuits = 'Tailored Suits',
  Tallow = 'Tallow',
  Tapestries = 'Tapestries',
  Teff = 'Teff',
  Telephones = 'Telephones',
  Timber = 'Timber',
  Tobacco = 'Tobacco',
  Tortillas = 'Tortillas',
  Toys = 'Toys',
  Typewriters = 'Typewriters',
  Violins = 'Violins',
  WanzaTimber = 'Wanza Timber',
  WaterDrop = 'Water Drop',
  Weapons = 'Weapons',
  WhaleOil = 'Whale Oil',
  Windows = 'Windows',
  Wood = 'Wood',
  WoodVeneers = 'Wood Veneers',
  Wool = 'Wool',
  WorkClothes = 'Work Clothes',
  Zinc = 'Zinc',
}

export enum Boost {
  None = 'None',
  Electricity = 'Electricity',
  TractorBarn = 'Tractor Barn',
  Fertiliser = 'Fertilizer',
  Silo = 'Silo',
}

export enum DepartmentOfLaborPolicy {
  None = 'None',
  LandReformAct = 'Land Reform Act',
  SkilledLaborAct = 'Skilled Labor Act',
  GalvanicGrantsAct = 'Galvanic Grants Act',
  FactoryInspectionsAct = 'Factory Inspections Act',
  UnionSubsidiesAct = 'Union Subsidies Act',
}

export enum Region {
  Unknown = 'Unknown',
  OldWorld = 'Old World',
  NewWorld = 'New World',
  CapeTrelawney = 'Cape Trelawney',
  Arctic = 'The Arctic',
  Enbesa = 'Enbesa',
}
// ...
export enum CulturalSet {
  Unknown = 'Unknown',
  BronzeAge = 'Bronze Age',
  OriginsOfMankind = 'Origins of Mankind',
  DomesticAnimals = 'Domestic Animals',
  OceanPredators = 'Ocean Predators',
  PolarCircle = 'Polar Circle',
  ProudSavannah = 'Proud Savannah',
  TeemingLakes = 'Teeming Lakes',
  Amazons = 'Amazons',
  Andean = 'Andean',
  Enchanted = 'Enchanted',
  NearEast = 'Near East',
  Subalpine = 'Subalpine',
}